+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added
 
 - Added `odood server healthcheck` command. One step to make Odood container-friendly.
+- Added `odood server wait-pg` command
+- Added `odood server run --wait-pg` option
 
 ### Changed